requirements.txt
# Core LlamaIndex packages
llama-index-core>=0.10.0
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-embeddings-voyageai
llama-index-readers-web
llama-index-vector-stores-chroma

# API integrations
google-cloud-language>=2.13.0
firecrawl-py>=1.0.0
openai>=1.0.0

# Data processing
pydantic>=2.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
markdown-it-py>=3.0.0
html5lib

# NLP and embeddings
spacy>=3.7.0
chromadb>=0.5.0  # Vector store for RAG testing

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0